refactor(middleware): reorganize imports and clean up unused constants
